requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Django==3.2
django-jazzmin==2.6.0
django-leaflet==0.28.3
Fiona==1.9.1
geopandas==0.12.2
munch==2.5.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
psycopg2-binary==2.9.5
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
shapely==2.0.1
six==1.16.0
sqlparse==0.4.3
